+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Added `withTokenBucket` helper function, which executes a callback with an appropriate
+  delay according to a provided `HierarchicalTokenBucket`.
+
 ## 0.1.0 - 2022-05-05
 
 ### Added